Corregidos algunas inconsistencias referidas al nombre de los paquetes
